+++ src/devices/tuya.ts
@@ -11,7 +11,8 @@ import {ColorMode, colorModeLookup} from '../lib/constants';
 import fz from '../converters/fromZigbee';
 import tz from '../converters/toZigbee';
 import {KeyValue, Definition, Tz, Fz, Expose, KeyValueAny, KeyValueString} from '../lib/types';
-import {onOff, quirkCheckinInterval, battery, deviceEndpoints, light, iasZoneAlarm, temperature, humidity, identify} from '../lib/modernExtend';
+import {onOff, quirkCheckinInterval, battery, deviceEndpoints, light, iasZoneAlarm, temperature, humidity, identify,
+    actionEnumLookup, commandsOnOff, commandsLevelCtrl} from '../lib/modernExtend';
 import {logger} from '../lib/logger';
 
 const NS = 'zhc:tuya';
@@ -703,56 +704,72 @@ const definitions: Definition[] = [
         model: 'F00MB00-04-1',
         vendor: 'FORIA',
         description: '4 scenes switch',
-        fromZigbee: [tuya.fz.datapoints],
-        toZigbee: [tuya.tz.datapoints],
-        configure: tuya.configureMagicPacket,
-        exposes: [e.action(['scene_1', 'scene_2', 'scene_3', 'scene_4'])],
-        meta: {
-            multiEndpoint: true,
-            tuyaDatapoints: [
-                [1, 'action', tuya.valueConverter.static('scene_1')],
-                [2, 'action', tuya.valueConverter.static('scene_2')],
-                [3, 'action', tuya.valueConverter.static('scene_3')],
-                [4, 'action', tuya.valueConverter.static('scene_4')],
-            ],
-        },
+        extend: [
+            tuya.modernExtend.tuyaMagicPacket(),
+            tuya.modernExtend.combineActions([
+                tuya.modernExtend.dpAction({dp: 1, lookup: {'scene_1': 0}}),
+                tuya.modernExtend.dpAction({dp: 2, lookup: {'scene_2': 0}}),
+                tuya.modernExtend.dpAction({dp: 3, lookup: {'scene_3': 0}}),
+                tuya.modernExtend.dpAction({dp: 4, lookup: {'scene_4': 0}}),
+            ]),
+            tuya.modernExtend.dpBinary({
+                name: 'vibration',
+                dp: 0x6c, type: tuya.dataTypes.enum,
+                valueOn: ['ON', 1],
+                valueOff: ['OFF', 0],
+                description: 'Enable vibration',
+            }),
+            tuya.modernExtend.dpBinary({
+                name: 'approach',
+                dp: 0x6b, type: tuya.dataTypes.enum,
+                valueOn: ['ON', 1],
+                valueOff: ['OFF', 0],
+                description: 'Enable approach detection',
+            }),
+            tuya.modernExtend.dpBinary({
+                name: 'illumination',
+                dp: 0x6a, type: tuya.dataTypes.enum,
+                valueOn: ['ON', 1],
+                valueOff: ['OFF', 0],
+                description: 'Enable illumination detection',
+            }),
+            tuya.modernExtend.dpBinary({
+                name: 'backlight',
+                dp: 0x69, type: tuya.dataTypes.enum,
+                valueOn: ['ON', 1],
+                valueOff: ['OFF', 0],
+                description: 'Enable backlight',
+            }),
+        ],
     },
     {
         fingerprint: tuya.fingerprint('TS0601', ['_TZE200_dhke3p9w']),
         model: 'F00YK04-18-1',
         vendor: 'FORIA',
         description: '18 scenes remote',
-        fromZigbee: [tuya.fz.datapoints],
-        toZigbee: [tuya.tz.datapoints],
-        configure: tuya.configureMagicPacket,
-        exposes: [e.action(['scene_1', 'scene_2', 'scene_3', 'scene_4', 'scene_5', 'scene_6',
-            'scene_7', 'scene_8', 'scene_9', 'scene_10', 'scene_11', 'scene_12', 'scene_13', 'scene_14', 'scene_15', 'scene_16',
-            'scene_17', 'scene_18']),
+        extend: [
+            tuya.modernExtend.tuyaMagicPacket(),
+            tuya.modernExtend.combineActions([
+                tuya.modernExtend.dpAction({dp: 1, lookup: {'scene_1': 0}}),
+                tuya.modernExtend.dpAction({dp: 2, lookup: {'scene_2': 0}}),
+                tuya.modernExtend.dpAction({dp: 3, lookup: {'scene_3': 0}}),
+                tuya.modernExtend.dpAction({dp: 4, lookup: {'scene_4': 0}}),
+                tuya.modernExtend.dpAction({dp: 5, lookup: {'scene_5': 0}}),
+                tuya.modernExtend.dpAction({dp: 6, lookup: {'scene_6': 0}}),
+                tuya.modernExtend.dpAction({dp: 7, lookup: {'scene_7': 0}}),
+                tuya.modernExtend.dpAction({dp: 8, lookup: {'scene_8': 0}}),
+                tuya.modernExtend.dpAction({dp: 9, lookup: {'scene_9': 0}}),
+                tuya.modernExtend.dpAction({dp: 10, lookup: {'scene_10': 0}}),
+                tuya.modernExtend.dpAction({dp: 11, lookup: {'scene_11': 0}}),
+                tuya.modernExtend.dpAction({dp: 12, lookup: {'scene_12': 0}}),
+                tuya.modernExtend.dpAction({dp: 13, lookup: {'scene_13': 0}}),
+                tuya.modernExtend.dpAction({dp: 14, lookup: {'scene_14': 0}}),
+                tuya.modernExtend.dpAction({dp: 15, lookup: {'scene_15': 0}}),
+                tuya.modernExtend.dpAction({dp: 16, lookup: {'scene_16': 0}}),
+                tuya.modernExtend.dpAction({dp: 101, lookup: {'scene_17': 0}}),
+                tuya.modernExtend.dpAction({dp: 102, lookup: {'scene_18': 0}}),
+            ]),
         ],
-
-        meta: {
-            multiEndpoint: true,
-            tuyaDatapoints: [
-                [1, 'action', tuya.valueConverter.static('scene_1')],
-                [2, 'action', tuya.valueConverter.static('scene_2')],
-                [3, 'action', tuya.valueConverter.static('scene_3')],
-                [4, 'action', tuya.valueConverter.static('scene_4')],
-                [5, 'action', tuya.valueConverter.static('scene_5')],
-                [6, 'action', tuya.valueConverter.static('scene_6')],
-                [7, 'action', tuya.valueConverter.static('scene_7')],
-                [8, 'action', tuya.valueConverter.static('scene_8')],
-                [9, 'action', tuya.valueConverter.static('scene_9')],
-                [10, 'action', tuya.valueConverter.static('scene_10')],
-                [11, 'action', tuya.valueConverter.static('scene_11')],
-                [12, 'action', tuya.valueConverter.static('scene_12')],
-                [13, 'action', tuya.valueConverter.static('scene_13')],
-                [14, 'action', tuya.valueConverter.static('scene_14')],
-                [15, 'action', tuya.valueConverter.static('scene_15')],
-                [16, 'action', tuya.valueConverter.static('scene_16')],
-                [101, 'action', tuya.valueConverter.static('scene_17')],
-                [102, 'action', tuya.valueConverter.static('scene_18')],
-            ],
-        },
     },
     {
         fingerprint: tuya.fingerprint('TS0601', [
@@ -6108,6 +6125,7 @@ const definitions: Definition[] = [
             tuya.exposes.powerFactorWithPhase('a'), tuya.exposes.powerFactorWithPhase('b'), tuya.exposes.powerFactorWithPhase('c'),
         ],
         meta: {
+            multiEndpointSkip: ['power_factor', 'power_factor_phase_b', 'power_factor_phase_c', 'energy'],
             tuyaDatapoints: [
                 [132, 'ac_frequency', tuya.valueConverter.raw],
                 [133, 'temperature', tuya.valueConverter.divideBy10],
@@ -7667,6 +7685,7 @@ const definitions: Definition[] = [
             e.numeric('update_frequency', ea.STATE).withUnit('s').withDescription('Update frequency'),
         ],
         meta: {
+            multiEndpointSkip: ['power_factor', 'power_factor_phase_b', 'power_factor_phase_c', 'energy'],
             tuyaDatapoints: [
                 [111, 'ac_frequency', tuya.valueConverter.divideBy100],
                 [101, 'power_a', tuya.valueConverter.divideBy10],
@@ -7705,6 +7724,7 @@ const definitions: Definition[] = [
             tuya.exposes.energyProducedWithPhase('a'), tuya.exposes.energyProducedWithPhase('b'),
         ],
         meta: {
+            multiEndpointSkip: ['power_factor', 'power_factor_phase_b', 'power_factor_phase_c', 'energy'],
             tuyaDatapoints: [
                 [113, 'ac_frequency', tuya.valueConverter.raw],
                 [108, 'power_a', tuya.valueConverter.raw],
@@ -8026,21 +8046,21 @@ const definitions: Definition[] = [
         model: 'F00XN00-04-1',
         vendor: 'FORIA',
         description: 'Dimmer 4 scenes',
-        fromZigbee: [fz.battery, fz.command_on, fz.command_off, fz.command_move_to_level, fz.command_move_to_color_temp,
-            fz.command_step_color_temperature, fz.command_step],
-        toZigbee: [],
-        exposes: [e.battery(), e.battery_voltage(), e.action(['on', 'off', 'brightness_move_to_level', 'color_temperature_move'])],
-        configure: async (device, coordinatorEndpoint) => {
-            const endpoint = device.getEndpoint(1);
-            await tuya.configureMagicPacket(device, coordinatorEndpoint);
-            await endpoint.command('genGroups', 'miboxerSetZones', {zones: [
-                {zoneNum: 1, groupId: 101},
-                {zoneNum: 2, groupId: 102},
-                {zoneNum: 3, groupId: 103},
-                {zoneNum: 4, groupId: 104},
-            ]});
-            await endpoint.command('genBasic', 'tuyaSetup', {}, {disableDefaultResponse: true});
-        },
+        extend: [
+            tuya.modernExtend.tuyaMagicPacket(),
+            battery({voltage: true}),
+            tuya.modernExtend.combineActions([
+                actionEnumLookup({
+                    actionLookup: {'scene_1': 1, 'scene_2': 2, 'scene_3': 3, 'scene_4': 4},
+                    cluster: 'genOnOff',
+                    commands: ['commandTuyaAction'],
+                    attribute: 'data',
+                    parse: (msg, attr) => msg.data[attr][1],
+                }),
+                commandsOnOff(),
+                commandsLevelCtrl({commands: ['brightness_move_up', 'brightness_move_down', 'brightness_stop']}),
+            ]),
+        ],
     },
     {
         fingerprint: tuya.fingerprint('TS0601', ['_TZE204_l6llgoxq']),
@@ -8112,6 +8132,23 @@ const definitions: Definition[] = [
                 [18, 'temperature', tuya.valueConverter.raw],
                 [19, 'humidity', tuya.valueConverter.raw],
             ],
+        },
+    },
+    {
+        fingerprint: tuya.fingerprint('TS110E', ['_TZ3210_guijtl8k']),
+        model: 'QS-Zigbee-D04',
+        vendor: 'LEDRON',
+        description: '0-10v dimmer',
+        fromZigbee: [fzLocal.TS110E, fz.on_off],
+        toZigbee: [tzLocal.TS110E_onoff_brightness, tzLocal.TS110E_options, tz.light_brightness_move],
+        exposes: [
+            e.light_brightness().withMinBrightness().withMaxBrightness(),
+        ],
+        configure: async (device, coordinatorEndpoint) => {
+            await tuya.configureMagicPacket(device, coordinatorEndpoint);
+            const endpoint = device.getEndpoint(1);
+            await reporting.bind(endpoint, coordinatorEndpoint, ['genOnOff', 'genLevelCtrl']);
+            await reporting.onOff(endpoint);
         },
     },
     {